+++ main/components/sensors/lightSensor.h
@@ -4,4 +4,4 @@
 #include "esp_log.h"
 
 int init_light_sensor(int ADC_CHANNEL);
-float get_light_value();
+int get_light_value();
